Move vectorToVariant to BaseVector
