+++ velox/vector/VariantToVector.cpp
@@ -263,109 +263,4 @@ VectorPtr variantToVector(
   return BaseVector::wrapInConstant(1, 0, std::move(variantVector));
 }
 
-namespace {
-
-Variant nullVariant(const TypePtr& type) {
-  return Variant(type->kind());
-}
-
-template <TypeKind kind>
-Variant variantAt(const VectorPtr& vector, int32_t row) {
-  using T = typename TypeTraits<kind>::NativeType;
-
-  const T value = vector->as<SimpleVector<T>>()->valueAt(row);
-
-  if (vector->type()->providesCustomComparison()) {
-    return Variant::typeWithCustomComparison<kind>(value, vector->type());
-  }
-
-  return Variant(value);
-}
-
-template <>
-Variant variantAt<TypeKind::VARBINARY>(const VectorPtr& vector, int32_t row) {
-  return Variant::binary(vector->as<SimpleVector<StringView>>()->valueAt(row));
-}
-
-Variant variantAt(const VectorPtr& vector, vector_size_t row);
-
-Variant arrayVariantAt(const VectorPtr& vector, vector_size_t row) {
-  auto arrayVector = vector->wrappedVector()->as<ArrayVector>();
-  auto& elements = arrayVector->elements();
-
-  auto wrappedRow = vector->wrappedIndex(row);
-  auto offset = arrayVector->offsetAt(wrappedRow);
-  auto size = arrayVector->sizeAt(wrappedRow);
-
-  std::vector<Variant> array;
-  array.reserve(size);
-  for (auto i = 0; i < size; i++) {
-    auto innerRow = offset + i;
-    array.push_back(variantAt(elements, innerRow));
-  }
-  return Variant::array(array);
-}
-
-Variant mapVariantAt(const VectorPtr& vector, vector_size_t row) {
-  auto mapVector = vector->wrappedVector()->as<MapVector>();
-  auto& mapKeys = mapVector->mapKeys();
-  auto& mapValues = mapVector->mapValues();
-
-  auto wrappedRow = vector->wrappedIndex(row);
-  auto offset = mapVector->offsetAt(wrappedRow);
-  auto size = mapVector->sizeAt(wrappedRow);
-
-  std::map<Variant, Variant> map;
-  for (auto i = 0; i < size; i++) {
-    auto innerRow = offset + i;
-    auto key = variantAt(mapKeys, innerRow);
-    auto value = variantAt(mapValues, innerRow);
-    map.insert({key, value});
-  }
-  return Variant::map(map);
-}
-
-Variant rowVariantAt(const VectorPtr& vector, vector_size_t row) {
-  auto rowValues = vector->wrappedVector()->as<RowVector>();
-  auto wrappedRow = vector->wrappedIndex(row);
-
-  std::vector<Variant> values;
-  for (auto& child : rowValues->children()) {
-    values.push_back(variantAt(child, wrappedRow));
-  }
-  return Variant::row(std::move(values));
-}
-
-Variant variantAt(const VectorPtr& vector, vector_size_t row) {
-  if (vector->isNullAt(row)) {
-    return nullVariant(vector->type());
-  }
-
-  auto typeKind = vector->typeKind();
-  if (typeKind == TypeKind::ROW) {
-    return rowVariantAt(vector, row);
-  }
-
-  if (typeKind == TypeKind::ARRAY) {
-    return arrayVariantAt(vector, row);
-  }
-
-  if (typeKind == TypeKind::MAP) {
-    return mapVariantAt(vector, row);
-  }
-
-  if (typeKind == TypeKind::OPAQUE) {
-    return Variant::opaque(
-        vector->as<SimpleVector<std::shared_ptr<void>>>()->valueAt(row),
-        std::dynamic_pointer_cast<const OpaqueType>(vector->type()));
-  }
-
-  return VELOX_DYNAMIC_SCALAR_TYPE_DISPATCH(variantAt, typeKind, vector, row);
-}
-} // namespace
-
-Variant vectorToVariant(const VectorPtr& vector, vector_size_t index) {
-  return variantAt(vector, index);
-}
-
 } // namespace facebook::velox
